EulerFlow: static cast to avoid compiler warnings.

+++ Code/ApplicationExamples/EulerFlow/MyEulerSolver.cpp
@@ -24,7 +24,7 @@ void Euler::MyEulerSolver::init(std::vector<std::string>& cmdlineargs) {
   // Demonstration how to access parameters:
   printf("EulerFlow was called with these parameters:\n");
   for(size_t i=0; i<cmdlineargs.size(); i++)
-    printf("%i. %s\n", i, cmdlineargs[i].c_str());
+    printf("%i. %s\n", (int)i, cmdlineargs[i].c_str());
 }
 
 void Euler::MyEulerSolver::flux(const double* const Q, double** F) {
